3rd Year/CS334/LocaliAPI/requirements.txt
django==3.0.6
djangorestframework==3.11.0
mysqlclient==1.4.6
mysql-connector-python==8.0.20
django-mysql==2.2.0
cryptography==2.9.2
Pillow==7.1.2
ordereddict==1.1
pymysql==0.9.3
django-jet==1.0.8
drf_yasg==1.17.1
django-filter==2.2.0